refactor(cli): migrate gee tooling onto discovery
Move GEE's refresher/writer/prober/emitter/coverage plus the credentialed hydrator
and the stanza categoriser to earthlens.gee.cli, relocating the _categories and
_hydrate helper modules into the provider distribution. Generalise core's
write_stanza categoriser and datasets --fill-empty hydrator to role dispatch (ecmwf
literals remain until it migrates); route the STAC-doc filename through
base.safe_filename. Fix core probe tests that had silently gone live to mock the
stac_cli seam. Shrink both guard allow-lists (gee).
