+++ libs/providers/imagery/src/earthlens/gee/_categories.py
